ml/requirements.txt
google-cloud-storage
firebase_admin
tensorflow>=2.1.0
pandas
scikit-learn